+++ IBAnimatable/AnimatorFactory.swift
@@ -18,11 +18,11 @@ public struct AnimatorFactory {
     case .systemRippleEffect:
       return SystemTransitionAnimator(systemType: .rippleEffect, duration: transitionDuration)
     case let .explode(xFactor, minAngle, maxAngle):
-      return ExplodeAnimator(xFactor: xFactor, minAngle: minAngle, maxAngle: maxAngle, transitionDuration: transitionDuration)
+      return ExplodeAnimator(xFactor: xFactor, minAngle: minAngle, maxAngle: maxAngle, duration: transitionDuration)
     case let .fade(direction):
-      return FadeAnimator(direction: direction, transitionDuration: transitionDuration)
+      return FadeAnimator(direction: direction, duration: transitionDuration)
     case let .fold(direction, folds):
-      return FoldAnimator(from: direction, folds: folds, transitionDuration: transitionDuration)
+      return FoldAnimator(from: direction, folds: folds, duration: transitionDuration)
     case let .portal(direction, zoomScale):
       return PortalAnimator(from: direction, zoomScale: zoomScale, duration: transitionDuration)
     case let .natGeo(direction):
@@ -58,7 +58,7 @@ public struct AnimatorFactory {
                                   transitionDuration: Duration = defaultPresentationDuration) -> AnimatedPresenting {
     switch presentationAnimationType {
     case let .cover(direction):
-      return CoverAnimator(from: direction, transitionDuration: transitionDuration)
+      return CoverAnimator(from: direction, duration: transitionDuration)
     case .zoom:
       return ZoomAnimator(duration: transitionDuration)
     case .dropDown: