+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file.
 - Replace all `SystemAnimator` classes with `SystemTransitionAnimator`
 [#427](https://github.com/IBAnimatable/IBAnimatable/pull/427) by [@SD10](https://github.com/sd10)
 - `PresentationDesignable` now supports `contextFrameForPresentation` which allow you to present a controller with a custom configuration over another instead of being in fullscreen. Imitates `UIModalPresentationStyle.currentContext`
+- Transition Animator classes initializer parameter changed from transitionDuration -> duration
+[#446](https://github.com/IBAnimatable/IBAnimatable/pull/446) by [@SD10](https://github.com/sd10)
 
 #### Enhancements
 - Conserve custom layer mask when using `Animatable*` instead of removing them